requirements.txt
tqdm
PyYAML
Pillow
torch==2.0.1+cu118
torchvision==0.15.2+cu118